refactor: Move pose table to pages subdirectory
