+++ src/components/index.js
@@ -1,6 +1,4 @@
 import { Nav, NavDetailed } from "./Nav"
 import HexapodPlot from "./HexapodPlot"
 import DimensionsWidget from "./DimensionsWidget"
-import PoseTable from "./PoseTable"
-import { AlertBox } from "./generic"
-export { PoseTable, Nav, NavDetailed, AlertBox, HexapodPlot, DimensionsWidget }
+export { Nav, NavDetailed, HexapodPlot, DimensionsWidget }
