+++ src/components/pages/InverseKinematicsPage.js
@@ -3,7 +3,7 @@ import { sliderList, Card, BasicButton, AlertBox } from "../generic"
 import { solveInverseKinematics } from "../../hexapod"
 import { SECTION_NAMES, IK_SLIDERS_LABELS, RESET_LABEL } from "../vars"
 import { DEFAULT_IK_PARAMS, DEFAULT_POSE } from "../../templates"
-import { PoseTable } from ".."
+import PoseTable from "./PoseTable"
 
 class InverseKinematicsPage extends Component {
     pageName = SECTION_NAMES.inverseKinematics
